+++ CHANGELOG.md
@@ -8,6 +8,7 @@ for commit guidelines.
 
 - Update build configuration and package exports
 - Update ESM target to ES2015 for backward compatibility
+- Add support for deleting a service.
 
 ## 0.2.6 (2024-09-02)